refactor: update the package to and deps
Package was previously ported from an existing site setup with some
minor changes. To bring the package more up to date and to ensure it
is suitable for further development, I have update the dependencies
included and ensured that the breaking changes have been resolved.

In addition to that, the package configs, plugins and rules are now
all seperated out from the main webpack object into their own files
where they can be used as when required.