fix: remove empty projects before processing
Fixes an issue where the lack of defined projects would cause standalone
builds to not occur due to the path not being found. This would happen
because the process would assume an empty entry in the projects list
was valid, so would attempt to treat it as such and seek through
defined plugin and theme directories. Standalone should no seek for
those directories.